move css to styles
