+++ src/pages/IndexPage.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import './App.css'
+import './styles/App.css'
 import Albedo from "./persons/Albedo";
 
 const IndexPage = () => {
