+++ src/pages/styles/AStarPage.css
@@ -166,7 +166,7 @@
     height: 91vh;
     align-items: center;
     align-content: center;
-    background: rgb(244, 244, 244) url("images/a-star-bg.png");
+    background: rgb(244, 244, 244) url("../images/a-star-bg.png");
 }
 
 #matrix_box{
