+++ src/pages/styles/App.css
@@ -54,7 +54,7 @@ a {
 }
 
 .menu {
-  background-image: url("images/background.jpg");
+  background-image: url("../images/background.jpg");
   background-color: #282c34;
   height: 100vh;
   display: flex;
