+++ src/pages/styles/GenPage.css
